Possible GPS Fix
It looks like we get correct data from the gps when using the gnss demo thing with this fix, however, we were only able to get time data from the gps. Guessing that a mudge dorm window doesn't get a lot of satellite coverage. It's disturbingly elegant and makes my work feel trivial. D;

+++ src/sensors/BuggySkeletonCode/BuggySkeletonCode.ino
@@ -9,7 +9,7 @@
 const int RF_PIN;
 //const int IMU_PIN;
 
-const int BAUD_RATE = 115200;
+const unsigned long BAUD_RATE = 115200;
 const int FREQUENCY = 10; //Hertz (must divide 1000 evenly)
 
 const byte LED = 13;
